Ignore .env
Ideally, we'd have `.env` in version control and use `.env.development` to
override local settings.

Unfortunately, `foreman` (or `heroku local`) won't use `.env.development` to
overwrite variables from `.env`. So if we have `API_KEY=ask_for_it` in `.env`
and the real API key in `.env.development`, it will be set to `ask_for_it` when
running with `heroku local`.

(Interestingly `rails server` works fine, possibly because it's single-threaded
while Foreman is multi-threaded.)

To fix the issue of not being able to override keys, switch back to an
old-skool `.env.example` which gets copied to `.env` in `bin/setup`.